+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/amazfitbip/AmazfitBipCooordinator.java
@@ -1,4 +1,4 @@
-/*  Copyright (C) 2017 Andreas Shimokawa
+/*  Copyright (C) 2017 Andreas Shimokawa, João Paulo Barraca
 
     This file is part of Gadgetbridge.
 
@@ -52,10 +52,9 @@ public class AmazfitBipCooordinator extends MiBand2Coordinator {
         return DeviceType.UNKNOWN;
     }
 
-
     @Override
     public InstallHandler findInstallHandler(Uri uri, Context context) {
-        return null; // not supported yet
+        AmazfitBipFWInstallHandler handler = new AmazfitBipFWInstallHandler(uri, context);
+        return handler.isValid() ? handler : null;
     }
-
 }