+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/btle/profiles/alertnotification/AlertCategory.java
@@ -1,4 +1,4 @@
-/*  Copyright (C) 2016-2017 Carsten Pfeiffer, Uwe Hermann
+/*  Copyright (C) 2016-2017 Andreas Shimokawa, Carsten Pfeiffer, Uwe Hermann
 
     This file is part of Gadgetbridge.
 
@@ -38,7 +38,7 @@ public enum AlertCategory {
     // 251-255 defined by service specification
     Any(255),
     Custom(-1),
-    CustomAmazfitBip(-6);
+    CustomMiBand2(-6);
 
     private final int id;