+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/btle/profiles/alertnotification/AlertNotificationProfile.java
@@ -1,4 +1,4 @@
-/*  Copyright (C) 2016-2017 Carsten Pfeiffer
+/*  Copyright (C) 2016-2017 Andreas Shimokawa, Carsten Pfeiffer
 
     This file is part of Gadgetbridge.
 
@@ -103,7 +103,7 @@ public class AlertNotificationProfile<T extends AbstractBTLEDeviceSupport> exten
         ByteArrayOutputStream stream = new ByteArrayOutputStream(100);
         stream.write(BLETypeConversions.fromUint8(alert.getCategory().getId()));
         stream.write(BLETypeConversions.fromUint8(alert.getNumAlerts()));
-        if (alert.getCategory() == AlertCategory.CustomAmazfitBip) {
+        if (alert.getCategory() == AlertCategory.CustomMiBand2) {
             stream.write(BLETypeConversions.fromUint8(alert.getCustomIcon()));
         }