+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/amazfitbip/AmazfitBipSupport.java
@@ -16,12 +16,16 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>. */
 package nodomain.freeyourgadget.gadgetbridge.service.devices.amazfitbip;
 
+import android.net.Uri;
+import android.widget.Toast;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
+import java.util.SimpleTimeZone;
 
 import nodomain.freeyourgadget.gadgetbridge.deviceevents.GBDeviceEventCallControl;
 import nodomain.freeyourgadget.gadgetbridge.devices.amazfitbip.AmazfitBipIcon;
@@ -35,9 +39,12 @@ import nodomain.freeyourgadget.gadgetbridge.service.btle.TransactionBuilder;
 import nodomain.freeyourgadget.gadgetbridge.service.btle.profiles.alertnotification.AlertCategory;
 import nodomain.freeyourgadget.gadgetbridge.service.btle.profiles.alertnotification.AlertNotificationProfile;
 import nodomain.freeyourgadget.gadgetbridge.service.btle.profiles.alertnotification.NewAlert;
+import nodomain.freeyourgadget.gadgetbridge.service.devices.amazfitbip.operations.AmazfitBipUpdateFirmwareOperation;
 import nodomain.freeyourgadget.gadgetbridge.service.devices.miband.NotificationStrategy;
 import nodomain.freeyourgadget.gadgetbridge.service.devices.miband2.MiBand2Support;
+import nodomain.freeyourgadget.gadgetbridge.util.GB;
 import nodomain.freeyourgadget.gadgetbridge.util.StringUtils;
+import nodomain.freeyourgadget.gadgetbridge.util.Version;
 
 public class AmazfitBipSupport extends MiBand2Support {
 
@@ -72,7 +79,7 @@ public class AmazfitBipSupport extends MiBand2Support {
 
             int customIconId = AmazfitBipIcon.mapToIconId(notificationSpec.type);
 
-            AlertCategory alertCategory = AlertCategory.CustomAmazfitBip;
+            AlertCategory alertCategory = AlertCategory.CustomMiBand2;
 
             // The SMS icon for AlertCategory.SMS is unique and not available as iconId
             if (notificationSpec.type == NotificationType.GENERIC_SMS) {
@@ -113,15 +120,40 @@ public class AmazfitBipSupport extends MiBand2Support {
     }
 
     @Override
+    public void onInstallApp(Uri uri) {
+        try {
+            new AmazfitBipUpdateFirmwareOperation(uri, this).perform();
+        } catch (IOException ex) {
+            GB.toast(getContext(), "Firmware cannot be installed: " + ex.getMessage(), Toast.LENGTH_LONG, GB.ERROR, ex);
+        }
+    }
+
+    @Override
     public void onSendWeather(WeatherSpec weatherSpec) {
         try {
             TransactionBuilder builder = performInitialized("Sending weather forecast");
+            Version version = new Version(gbDevice.getFirmwareVersion());
+
+            boolean supportsConditionString = false;
+            if (version.compareTo(new Version("0.0.8.74")) >= 0) {
+                supportsConditionString = true;
+            }
+
             final byte NR_DAYS = 2;
-            ByteBuffer buf = ByteBuffer.allocate(7 + 4 * NR_DAYS);
+            int bytesPerDay = 4;
+            int conditionsLength = 0;
+            if (supportsConditionString) {
+                bytesPerDay = 5;
+                conditionsLength = weatherSpec.currentCondition.getBytes().length;
+            }
+            int length = 7 + bytesPerDay * NR_DAYS + conditionsLength;
+            ByteBuffer buf = ByteBuffer.allocate(length);
+
             buf.order(ByteOrder.LITTLE_ENDIAN);
             buf.put((byte) 1);
             buf.putInt(weatherSpec.timestamp);
-            buf.put((byte) 0);
+            int tz_offset_hours = SimpleTimeZone.getDefault().getOffset(weatherSpec.timestamp * 1000L) / (1000 * 60 * 60);
+            buf.put((byte) (tz_offset_hours * 4));
 
             buf.put(NR_DAYS);
 
@@ -130,13 +162,19 @@ public class AmazfitBipSupport extends MiBand2Support {
             buf.put(condition);
             buf.put((byte) (weatherSpec.todayMaxTemp - 273));
             buf.put((byte) (weatherSpec.todayMinTemp - 273));
-
+            if (supportsConditionString) {
+                buf.put(weatherSpec.currentCondition.getBytes());
+                buf.put((byte) 0); //
+            }
             condition = AmazfitBipWeatherConditions.mapToAmazfitBipWeatherCode(weatherSpec.tomorrowConditionCode);
 
             buf.put(condition);
             buf.put(condition);
             buf.put((byte) (weatherSpec.tomorrowMaxTemp - 273));
             buf.put((byte) (weatherSpec.tomorrowMinTemp - 273));
+            if (supportsConditionString) {
+                buf.put((byte) 0); // not yet in weatherspec
+            }
 
             builder.write(getCharacteristic(AmazfitBipService.UUID_CHARACTERISTIC_WEATHER), buf.array());
             builder.queue(getQueue());