+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/miband2/operations/UpdateFirmwareOperation.java
@@ -1,4 +1,4 @@
-/*  Copyright (C) 2016-2017 Carsten Pfeiffer
+/*  Copyright (C) 2016-2017 Andreas Shimokawa, Carsten Pfeiffer
 
     This file is part of Gadgetbridge.
 
@@ -48,11 +48,11 @@ import nodomain.freeyourgadget.gadgetbridge.util.Prefs;
 public class UpdateFirmwareOperation extends AbstractMiBand2Operation {
     private static final Logger LOG = LoggerFactory.getLogger(UpdateFirmwareOperation.class);
 
-    private final Uri uri;
-    private final BluetoothGattCharacteristic fwCControlChar;
-    private final BluetoothGattCharacteristic fwCDataChar;
-    final Prefs prefs = GBApplication.getPrefs();
-    private Mi2FirmwareInfo firmwareInfo;
+    protected final Uri uri;
+    protected final BluetoothGattCharacteristic fwCControlChar;
+    protected final BluetoothGattCharacteristic fwCDataChar;
+    protected final Prefs prefs = GBApplication.getPrefs();
+    protected Mi2FirmwareInfo firmwareInfo;
 
     public UpdateFirmwareOperation(Uri uri, MiBand2Support support) {
         super(support);
@@ -82,7 +82,7 @@ public class UpdateFirmwareOperation extends AbstractMiBand2Operation {
         //the firmware will be sent by the notification listener if the band confirms that the metadata are ok.
     }
 
-    private void done() {
+    protected void done() {
         LOG.info("Operation done.");
         operationFinished();
         unsetBusy();
@@ -163,7 +163,7 @@ public class UpdateFirmwareOperation extends AbstractMiBand2Operation {
             done();
         }
     }
-    private void displayMessage(Context context, String message, int duration, int severity) {
+    protected void displayMessage(Context context, String message, int duration, int severity) {
         getSupport().handleGBDeviceEvent(new GBDeviceEventDisplayMessage(message, duration, severity));
     }